requirements.txt
websocket-server==0.6.4
pynput==1.8.1
mss==10.0.0
numpy==2.2.3
Pillow==10.3.0
opencv-python==4.11.0.86